Update LICENSE
There is no intended impact on licensing with this change. Somehow the copy of the BSD-3-clause license had typos/drift in it, which is completely unintentional. I might have been copying that from project to project over many years. This restores it to the BSD-3-clause text from OSI.